Move printers to core test module
