+++ java/core/src/test/java/sleeper/core/testutils/printers/FileInfoPrinterTest.shouldPrintDifferentFilesForOneTable.approved.txt
