+++ java/core/src/test/java/sleeper/core/testutils/printers/FileInfoPrinterTest.shouldPrintFilesOnLeaves.approved.txt
