+++ java/core/src/test/java/sleeper/core/testutils/printers/FileInfoPrinterTest.shouldPrintFilesOnceWhenTwoTablesAreIdentical.approved.txt
