+++ java/core/src/test/java/sleeper/core/testutils/printers/FileInfoPrinterTest.shouldPrintOnlyOneTable.approved.txt
