+++ java/core/src/test/java/sleeper/core/testutils/printers/PartitionsPrinterTest.shouldPrintDifferentPartitionsForOneTable.approved.txt
