+++ java/core/src/test/java/sleeper/core/testutils/printers/PartitionsPrinterTest.shouldRenamePartitionsByLocation.approved.txt
