+++ java/core/src/test/java/sleeper/core/testutils/printers/TablesPrinter.java
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.suite.testutil;
-
-import sleeper.configuration.properties.format.ToStringPrintStream;
+package sleeper.core.testutils.printers;
 
 import java.io.PrintStream;
 import java.util.Collection;
